requirements.txt
python-telegram-bot==13.17
pandas
requests
matplotlib